Implement fscroll and partially fix redisplay
Because vscroll does not correct for displaying images of different size, this
commit introduces fscroll, which represents the vscroll in units of current page
height.

Furthermore, the commit adds a mechanism to redisplay all buffer windows,
including scrolling to the correct position. Although the mechanism seems to
consistently set the correct page, for some reason it does not always set the
correct vscroll/fscroll.